10188: Groundwork to allow editing of mock user credentials

+++ app/src/main/java/com/pavilionpay/igaming/remote/ExistingPatronRequestDto.kt
@@ -4,16 +4,16 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ExistingPatronRequestDto(
-    val patronType: String,
-    val patronID: String,
-    val vipCardNumber: String,
-    val dateOfBirth: String,
-    val remainingDailyDeposit: Double,
-    val walletBalance: Double,
-    val transactionID: String,
-    val transactionAmount: Float,
-    val transactionType: Byte,
-    val returnURL: String,
-    val productType: String,
-    val androidPackageName: String,
+        val patronType: String,
+        val patronID: String,
+        val vipCardNumber: String,
+        val dateOfBirth: String,
+        val remainingDailyDeposit: Double,
+        val walletBalance: Double,
+        val transactionID: String,
+        val transactionAmount: Double,
+        val transactionType: Byte,
+        val returnURL: String,
+        val productType: String,
+        val androidPackageName: String,
 )
